// src/lib/api/axios/types.ts
export interface ResGetBuildHistoryList {
  id: string;
  status: BuildStatus;
  name: string;
}